Update plex-installer.sh; Plex Pass and Autoupdate support
I could not find any guidelines for contribution so I apologize in advance for any issues.

* Added support for choosing between Public or Plex Pass installation of Plex Media Server
* Added option to run plexupdater via a cronjob; daily, weekly, monthly, or manual.

I am not entirely sure about you color use for messages - they, therefor, might be wrong!

The scriptpart for the input of email and password for Plex Pass may seem a little complex in its construction. What it does it that it hides the input behind asterix stars. I chose the approach as emails are usually longer than passwords and I suspect that this will make it easier for users to correct mistakes (without having to retype everything).