Removed check for ISSETUP
The check for ISSETUP will not work if setup is changed, symlinked, or an alias. Rather then adding a bunch of logic to check that the file executed is setup.sh I set ISSETUP to YES. This should provide the same function of preventing direct access to the scripts in the /inc directory.