shellcheck
The surrounding quotes actually unquote this. Remove or escape them.
https://github.com/koalaman/shellcheck/wiki/SC2027

Use -print0/-0 or find -exec + to allow for non-alphanumeric filenames.
https://github.com/koalaman/shellcheck/wiki/SC2038

Quote this to prevent word splitting
https://github.com/koalaman/shellcheck/wiki/SC2046

Tips depend on target shell and yours is unknown. Add a shebang.
https://github.com/koalaman/shellcheck/wiki/SC2148

Possible Misspelling: MYVARIABLE may not be assigned, but MY_VARIABLE
is.
https://github.com/koalaman/shellcheck/wiki/SC2153

Declare and assign separately to avoid masking return values.
https://github.com/koalaman/shellcheck/wiki/SC2155

Use cd ... || exit in case cd fails.
https://github.com/koalaman/shellcheck/wiki/SC2164

Prefer [ p ] && [ q ] as [ p -a q ] is not well defined.
https://github.com/koalaman/shellcheck/wiki/SC2166